move submission/clinicaData from routing into Views
