+++ src/app/(post-login)/submission/program/[shortName]/clinical-data/page.tsx
@@ -18,285 +18,8 @@
  */
 'use client';
 
-import { ClinicalEntitySearchResultsQuery } from '@/__generated__/clinical/graphql';
-import ContentMain from '@/app/components/Content/ContentMain';
 import { pageWithPermissions } from '@/app/components/Page';
-import { BreadcrumbTitle, HelpLink, PageHeader } from '@/app/components/PageHeader/PageHeader';
-import CLINICAL_ENTITY_SEARCH_RESULTS_QUERY from '@/app/gql/clinical/CLINICAL_ENTITY_SEARCH_RESULTS_QUERY';
-import SUBMITTED_DATA_SIDE_MENU_QUERY from '@/app/gql/clinical/SUBMITTED_DATA_SIDE_MENU_QUERY';
-import { useClinicalQuery, useUrlParamState } from '@/app/hooks';
-import { notNull, parseDonorIdString } from '@/global/utils';
-import { css } from '@/lib/emotion';
-import { Container, Loader, Typography, VerticalTabs } from '@icgc-argo/uikit';
-import { useState } from 'react';
-import { setConfiguration } from 'react-grid-system';
-import ClinicalEntityDataTable from './ClinicalEntityDataTable';
-import ClinicalDownloadButton from './DownloadButtons';
-import SearchBar from './SearchBar';
-import {
-	ClinicalEntitySearchResultResponse,
-	CompletionStates,
-	TsvDownloadIds,
-	aliasedEntityNames,
-	clinicalEntityDisplayNames,
-	clinicalEntityFields,
-	defaultClinicalEntityFilters,
-	emptyClinicalDataResponse,
-	emptySearchResponse,
-	hasClinicalErrors,
-	reverseLookUpEntityAlias,
-} from './common';
-
-setConfiguration({ gutterWidth: 9 });
-
-const defaultClinicalEntityTab = aliasedEntityNames.donor;
-
-// convert codegen automated types to definitions in existing code
-// make data object shape uniform
-const parseSearchResult = (
-	data: ClinicalEntitySearchResultsQuery | undefined,
-): ClinicalEntitySearchResultResponse => {
-	if (!data) return emptySearchResponse;
-	return {
-		clinicalSearchResults: {
-			searchResults: data.clinicalSearchResults.searchResults.map((searchResults) => ({
-				donorId: searchResults?.donorId || 0,
-				submitterDonorId: searchResults?.submitterDonorId || '',
-			})),
-			totalResults: data.clinicalSearchResults.totalResults,
-		},
-	};
-};
-
-const ClinicalDataPageComp = ({ programShortName }: { programShortName: string }) => {
-	const [keyword, setKeyword] = useState('');
-	const [completionState, setCompletionState] = useState(CompletionStates['all']);
-	const [modalVisible, setModalVisible] = useState(false);
-
-	const [selectedClinicalEntityTab, setSelectedClinicalEntityTab] = useUrlParamState(
-		'tab',
-		defaultClinicalEntityTab,
-	);
-	const [selectedDonors, setSelectedDonors] = useUrlParamState('donorId', '');
-
-	const currentEntity = reverseLookUpEntityAlias(selectedClinicalEntityTab);
-	const urlDonorQueryStrings = selectedDonors ? selectedDonors.split(',') : [];
-	const currentDonors = urlDonorQueryStrings.length
-		? urlDonorQueryStrings.map((donorId) => parseDonorIdString(donorId))
-		: [];
-
-	// Matches multiple digits and/or digits preceded by DO, followed by a comma, space, or end of string
-	// Example: DO259138, 2579137, DASH-7, DO253290abcdef
-	// Regex will match first 2 Donor IDs, but not 3rd Submitter ID or 4th case w/ random text
-	const searchDonorIds =
-		keyword
-			.match(/(?=DO|\d)\d+(?=,| |$)/gi)
-			?.filter((match) => !!match)
-			.map((idString) => parseInt(idString)) || [];
-
-	// Matches 'D' or 'DO' exactly (case insensitive)
-	const donorPrefixSearch = keyword.match(/^(d|do)\b/gi);
-
-	const searchSubmitterIds = donorPrefixSearch
-		? []
-		: keyword.split(/, |,/).filter((word) => !!word);
-
-	// Search Result Query
-	// Populates dropdown menu; Search query populates data table if there are no URL params
-	const { data: searchResultData, loading: searchResultsLoading } = useClinicalQuery(
-		CLINICAL_ENTITY_SEARCH_RESULTS_QUERY,
-		{
-			errorPolicy: 'all',
-			variables: {
-				programShortName,
-				filters: {
-					...defaultClinicalEntityFilters,
-					completionState,
-					entityTypes: [currentEntity],
-					donorIds: searchDonorIds,
-					submitterDonorIds: searchSubmitterIds,
-				},
-			},
-		},
-	);
-
-	const parsedSearchResultData = parseSearchResult(searchResultData);
-
-	const searchResults = parsedSearchResultData?.clinicalSearchResults?.searchResults;
-	const searchResultIds = searchResults.map((result) => result?.donorId).filter(notNull);
-
-	const sideMenuQueryDonorIds =
-		urlDonorQueryStrings.length > 0
-			? currentDonors
-			: searchResults.length > 0 && keyword.length > 0
-			? searchResultIds
-			: [];
-
-	// Side Menu Query
-	// Populates Clinical Entity Table, Side Menu, Title Bar
-	const { data: sideMenuQuery, loading: sideMenuLoading } = useClinicalQuery(
-		SUBMITTED_DATA_SIDE_MENU_QUERY,
-		{
-			errorPolicy: 'all',
-			variables: {
-				programShortName,
-				filters: {
-					...defaultClinicalEntityFilters,
-					completionState,
-					donorIds: sideMenuQueryDonorIds,
-				},
-			},
-		},
-	);
-
-	const sideMenuData =
-		sideMenuQuery == undefined || sideMenuLoading
-			? emptyClinicalDataResponse(programShortName)
-			: sideMenuQuery;
-
-	const { clinicalData } = sideMenuData;
-
-	const useDefaultQuery =
-		currentDonors.length === 0 &&
-		(donorPrefixSearch || (searchDonorIds.length === 0 && searchSubmitterIds.length === 0)) &&
-		completionState === 'all';
-
-	const noSearchData = searchResultData === null || searchResultData === undefined;
-	const noData =
-		clinicalData.clinicalEntities.length === 0 && noSearchData && currentDonors.length === 0;
-
-	const entityTableDonorIds =
-		currentDonors.length > 0 ? currentDonors : searchResults.length > 0 ? searchResultIds : [];
-
-	const entityTableSubmitterDonorIds = (searchResults || [])
-		.map((searchResults) => searchResults?.submitterDonorId)
-		.filter(Boolean);
-
-	const tsvDownloadIds: TsvDownloadIds = {
-		donorIds: useDefaultQuery ? [] : entityTableDonorIds,
-		submitterDonorIds: useDefaultQuery ? [] : entityTableSubmitterDonorIds.filter(notNull),
-	};
-
-	const menuItems = clinicalEntityFields.map((entity) => (
-		<VerticalTabs.Item
-			key={entity}
-			active={selectedClinicalEntityTab === aliasedEntityNames[entity]}
-			onClick={() => setSelectedClinicalEntityTab(aliasedEntityNames[entity])}
-			disabled={
-				!clinicalData.clinicalEntities.some((e) => e?.entityName === aliasedEntityNames[entity])
-			}
-		>
-			{clinicalEntityDisplayNames[entity]}
-			{hasClinicalErrors(clinicalData, entity) && (
-				<VerticalTabs.Tag variant="ERROR">!</VerticalTabs.Tag>
-			)}
-		</VerticalTabs.Item>
-	));
-
-	return (
-		<div>
-			<PageHeader
-				leftSlot={<BreadcrumbTitle breadcrumbs={[programShortName, 'Submitted Data']} />}
-				rightSlot={<HelpLink url="" />}
-			/>
-			{searchResultsLoading ? (
-				<Loader />
-			) : (
-				<>
-					<ContentMain>
-						<SearchBar
-							setModalVisible={setModalVisible}
-							modalVisible={modalVisible}
-							completionState={completionState}
-							setCompletionState={setCompletionState}
-							programShortName={programShortName}
-							keyword={keyword}
-							loading={searchResultsLoading}
-							noData={noData}
-							useDefaultQuery={useDefaultQuery}
-							currentDonors={currentDonors}
-							setSelectedDonors={setSelectedDonors}
-							tsvDownloadIds={tsvDownloadIds}
-							donorSearchResults={parsedSearchResultData}
-							setKeyword={setKeyword}
-						/>
-						<Container>
-							<div
-								css={css`
-									width: 100%;
-								`}
-							>
-								{/* Sidebar */}
-								<div
-									css={css`
-										width: 20%;
-										max-width: 170px;
-										display: inline-block;
-									`}
-								>
-									<VerticalTabs>{menuItems}</VerticalTabs>
-								</div>
-								{/* Content */}
-								<div
-									css={css`
-										display: inline-block;
-										height: 100%;
-										width: calc(97% - 170px);
-										vertical-align: top;
-										padding: 8px 12px;
-									`}
-								>
-									{/* Header */}
-									<div
-										css={css`
-											width: 100%;
-											display: flex;
-											justify-content: space-between;
-										`}
-									>
-										<Typography
-											variant="subtitle2"
-											css={css`
-												margin-top: 4px;
-												margin-left: 4px;
-											`}
-										>
-											{clinicalEntityDisplayNames[currentEntity]} Data
-										</Typography>
-
-										<ClinicalDownloadButton
-											tsvDownloadIds={tsvDownloadIds}
-											text={`${clinicalEntityDisplayNames[currentEntity]} Data`}
-											entityTypes={[currentEntity]}
-											completionState={completionState}
-											disabled={noData}
-										/>
-									</div>
-									{/* DataTable */}
-									<div
-										css={css`
-											margin-top: 16px;
-										`}
-									>
-										<ClinicalEntityDataTable
-											entityType={currentEntity}
-											program={programShortName}
-											completionState={completionState}
-											currentDonors={entityTableDonorIds}
-											donorSearchResults={searchResultData}
-											useDefaultQuery={useDefaultQuery}
-											noData={noData}
-										/>
-									</div>
-								</div>{' '}
-							</div>
-						</Container>
-					</ContentMain>
-				</>
-			)}
-		</div>
-	);
-};
+import ClinicalDataPageComp from '@/views/submission/clinical-data/ClinicalData';
 
 const ClinicalDataPage = ({ params: { shortName = '' } }: { params: { shortName: string } }) => {
 	const ClinicalDataWithPermissions = pageWithPermissions(ClinicalDataPageComp, {
