+++ src/views/submission/clinical-data/ClinicalDataTableComp/index.tsx
