+++ src/views/submission/clinical-data/ClinicalEntityDataTable/components.tsx
