+++ src/views/submission/clinical-data/ClinicalEntityDataTable/dataQuery.tsx
