+++ src/views/submission/clinical-data/ClinicalEntityDataTable/index.tsx
