+++ src/views/submission/clinical-data/ClinicalEntityDataTable/tableConfig.tsx
