+++ src/views/submission/clinical-data/ClinicalEntityDataTable/types.ts
