+++ src/views/submission/clinical-data/ClinicalEntityDataTable/util.ts
