+++ src/views/submission/clinical-data/SearchBar/FilterModal/index.tsx
