+++ src/views/submission/clinical-data/SearchBar/SearchResultsMenu.tsx
