+++ src/views/submission/clinical-data/SearchBar/index.tsx
