+++ src/views/submission/clinical-data/SearchBar/style.tsx
