+++ src/views/submission/clinical-data/common.tsx
